+++ src/main/java/com/plivo/api/models/verify_session/VerifySessionList.java
@@ -1,0 +1,8 @@
+package com.plivo.api.models.verify_session;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
+public class VerifySessionList extends VerifySession{
+  @JsonIgnore
+  private String apiId;
+}